Change: move modify_note to dedicated files
