+++ src/manage_notes.h
@@ -22,4 +22,9 @@ delete_note (const char *, int);
 int
 note_uuid (note_t, char **);
 
+int
+modify_note (const gchar *, const char *, const char *, const char *,
+             const char *, const char *, const char *, const char *,
+             const gchar *, const gchar *);
+
 #endif /* not _GVMD_MANAGE_NOTES_H */
