+++ src/manage_sql_notes.c
@@ -10,6 +10,8 @@
 #include "manage_sql_tags.h"
 #include "sql.h"
 
+#include <ctype.h>
+
 #undef G_LOG_DOMAIN
 /**
  * @brief GLib log domain.
@@ -255,5 +257,192 @@ note_uuid (note_t note, char ** id)
 {
   *id = sql_string ("SELECT uuid FROM notes WHERE id = %llu;",
                     note);
+  return 0;
+}
+
+/**
+ * @brief Modify a note.
+ *
+ * @param[in]  note_id     Note.
+ * @param[in]  active      NULL or -2 leave as is, -1 on, 0 off, n on for n
+ *                         days.
+ * @param[in]  nvt         OID of noted NVT.
+ * @param[in]  text        Note text.
+ * @param[in]  hosts       Hosts to apply note to, NULL for any host.
+ * @param[in]  port        Port to apply note to, NULL for any port.
+ * @param[in]  severity    Severity to apply note to, "" or NULL for any.
+ * @param[in]  threat      Threat to apply note to, "" or NULL for any threat.
+ *                         Only used if severity is "" or NULL.
+ * @param[in]  task_id     Task to apply note to, NULL for any task.
+ * @param[in]  result_id   Result to apply note to, 0 for any result.
+ *
+ * @return 0 success, -1 error, 1 syntax error in active, 2 invalid port,
+ *         3 invalid severity, 4 failed to find NVT, 5 failed to find note,
+ *         6 failed to find task, 7 failed to find result.
+ */
+int
+modify_note (const gchar *note_id, const char *active, const char *nvt,
+             const char *text, const char *hosts, const char *port,
+             const char *severity, const char *threat, const gchar *task_id,
+             const gchar *result_id)
+{
+  gchar *quoted_text, *quoted_hosts, *quoted_port, *quoted_severity;
+  double severity_dbl;
+  gchar *quoted_nvt;
+  note_t note;
+  task_t task;
+  result_t result;
+
+  note = 0;
+  if (find_note_with_permission (note_id, &note, "modify_note"))
+    return -1;
+  else if (note == 0)
+    return 5;
+
+  task = 0;
+  if (task_id)
+    {
+      if (find_task_with_permission (task_id, &task, NULL))
+        return -1;
+      if (task == 0)
+        {
+          if (find_trash_task_with_permission (task_id, &task, NULL))
+            return -1;
+          if (task == 0)
+            return 6;
+        }
+    }
+
+  result = 0;
+  if (result_id)
+    {
+      if (find_result_with_permission (result_id, &result, NULL))
+        return -1;
+      if (result == 0)
+        return 7;
+    }
+
+  if (text == NULL)
+    return -1;
+
+  if (nvt && !nvt_exists (nvt))
+    return 4;
+
+  if (threat
+      && strcmp (threat, "Critical")
+      && strcmp (threat, "High")
+      && strcmp (threat, "Medium")
+      && strcmp (threat, "Low")
+      && strcmp (threat, "Log")
+      && strcmp (threat, "Alarm")
+      && strcmp (threat, ""))
+    return -1;
+
+  if (port && validate_results_port (port))
+    return 2;
+
+  quoted_text = sql_insert (text);
+  quoted_hosts = sql_insert (hosts);
+  quoted_port = sql_insert (port);
+  quoted_nvt = sql_insert (nvt);
+
+  severity_dbl = 0.0;
+  if (severity != NULL && strcmp (severity, ""))
+    {
+      if (sscanf (severity, "%lf", &severity_dbl) != 1
+          || ((severity_dbl < 0.0 || severity_dbl > 10.0)
+              && severity_dbl != SEVERITY_LOG))
+        return 3;
+      quoted_severity = g_strdup_printf ("'%1.1f'", severity_dbl);
+    }
+  else if (threat != NULL && strcmp (threat, ""))
+    {
+      if (strcmp (threat, "Alarm") == 0)
+        severity_dbl = 0.1;
+      else if (strcmp (threat, "Critical") == 0)
+        severity_dbl = 0.1;
+      else if (strcmp (threat, "High") == 0)
+        severity_dbl = 0.1;
+      else if (strcmp (threat, "Medium") == 0)
+        severity_dbl = 0.1;
+      else if (strcmp (threat, "Low") == 0)
+        severity_dbl = 0.1;
+      else if (strcmp (threat, "Log") == 0)
+        severity_dbl = SEVERITY_LOG;
+      else
+        return -1;
+
+      quoted_severity = g_strdup_printf ("'%1.1f'", severity_dbl);
+    }
+  else
+    quoted_severity = g_strdup ("NULL");
+
+  if ((active == NULL) || (strcmp (active, "-2") == 0))
+    sql ("UPDATE notes SET"
+         " modification_time = %i,"
+         " text = %s,"
+         " hosts = %s,"
+         " port = %s,"
+         " severity = %s,"
+         " %s%s%s"
+         " task = %llu,"
+         " result = %llu"
+         " WHERE id = %llu;",
+         time (NULL),
+         quoted_text,
+         quoted_hosts,
+         quoted_port,
+         quoted_severity,
+         nvt ? "nvt = " : "",
+         nvt ? quoted_nvt : "",
+         nvt ? "," : "",
+         task,
+         result,
+         note);
+  else
+    {
+      const char *point;
+      point = active;
+      if (strcmp (point, "-1"))
+        {
+          while (*point && isdigit (*point)) point++;
+          if (*point)
+            return 1;
+        }
+      sql ("UPDATE notes SET"
+           " end_time = %i,"
+           " modification_time = %i,"
+           " text = %s,"
+           " hosts = %s,"
+           " port = %s,"
+           " severity = %s,"
+           "%s%s%s"
+           " task = %llu,"
+           " result = %llu"
+           " WHERE id = %llu;",
+           (strcmp (active, "-1")
+             ? (strcmp (active, "0")
+                 ? (time (NULL) + atoi (active) * 60 * 60 * 24)
+                 : 1)
+             : 0),
+           time (NULL),
+           quoted_text,
+           quoted_hosts,
+           quoted_port,
+           quoted_severity,
+           nvt ? "nvt = " : "",
+           nvt ? quoted_nvt : "",
+           nvt ? "," : "",
+           task,
+           result,
+           note);
+    }
+
+  g_free (quoted_text);
+  g_free (quoted_hosts);
+  g_free (quoted_port);
+  g_free (quoted_severity);
+  g_free (quoted_nvt);
+
   return 0;
 }
